refactor: implemented requested changes
